restart recovery, settings panel updates 1. some errors will now automatically restart to settings panel 2. fixed some bugs with the settings panel, setting wifi credentials and online go credentials now works correctly

+++ firmware/go_board_firmware_std/src/onlinego/api.rs
@@ -308,7 +308,7 @@ impl BoardState {
             .iter()
             .enumerate()
             .map(|(x, v)| (x, v.iter().enumerate()))
-            .flat_map(|(x, iter)| iter.map(move |(y, v)| (x, y, v)))
+            .flat_map(|(x, iter)| iter.map(move |(y, v)| (x as u8, y as u8, v)))
     }
 }
